requirements.txt
ftfy
packaging
regex
tqdm
torch
torchvision
gradio
matplotlib
pillow
numpy
git+https://github.com/openai/CLIP.git
